Changed dependencies so the structure is somewhat modular
