+++ Backend.hpp
@@ -1,9 +1,16 @@
+#include <string>
 #include <iostream>
 #include <SFML/Window.hpp>
-#include "UI.hpp"
 #include "macros.hpp"
 #include "globalVars.hpp"
 
+using namespace std;
+
 void createBoard();
 
-void linkDots(sf::Event::MouseButtonEvent click1);
+int getSizeFromStr(string s);
+
+bool isClickValid(sf::Event::MouseButtonEvent mouse);
+
+void linkIfValid(sf::Event::MouseButtonEvent click1, sf::Event::MouseButtonEvent click2);
+
