+++ main.cpp
@@ -6,37 +6,13 @@ using namespace std;
 
 using namespace sf;
 
-void StartGame()
-{
-    createBoard();
-    window.create(sf::VideoMode((boardSize - 1) * colDist + 2 * circleRadius, (boardSize - 1) * rowDist + 2 * circleRadius), "Bridg-It", Style::Titlebar | Style::Close);
-    while (window.isOpen())
-    {
-        sf::Event event;
-        while (window.pollEvent(event))
-        {
-            if (event.type == sf::Event::Closed)
-                window.close();
-            if (event.type == Event::MouseButtonPressed){
-                if(event.mouseButton.button == Mouse::Left){
-                    linkDots(event.mouseButton);
-                }
-            }
-        }
 
-        window.clear();
-        loadBoard();
-        window.display();
-    }
-
-
-}
 
 
 
 int main()
 {
-    //StartGame();
+    startGame();
     //window.create(sf::VideoMode(800,600), "Bridg-It");
     gameOptionsMenu();
     //Meniusetup();
